IO#widen uses Liskov, not Scala variance.
This came up on gitter. The goal is for `IO[Void, A]` to be able
to be widened to `IO[E, A]` for any `A`. `Void` is abstract and
upper bounded by `Any`, so Scala doesn't let anything else be
passed into `widen`'s type argument. Instead of using `>:`, just
use `Liskov`.